requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
librosa==0.10.1
soundfile==0.12.1
SpeechRecognition==3.10.0
pydub==0.25.1
scikit-learn==1.3.0
pytest==7.4.2
black==23.9.1
flake8==6.1.0

# Optional: PyTorch and Transformers (for advanced emotion detection)
# Uncomment the lines below if you want to use transformer models
# torch==2.0.1
# transformers==4.35.0 